SNOW-1980772: PR cleanup
